Move `MapsView` class into dedicated module
