+++ metrics/api/views/maps/__init__.py
@@ -1,0 +1,1 @@
+from .api_view import MapsView
